First commit for djflipmix.com with all base files.
Beginning project by setting up base files in css and javascript.
Bootstrap 3 will be used for this project and will updated with
Bootstrap 4 later on.

Created a blank html file with references to all relevant files. A
template may be used as client has suggested one as an example of how he
wants his site to behave like. If template is used completely or for
inspiration, it will be documented.

The icons will be from font-awesome for easy scaling. I requested some
images from the client that will be used until I either receive higher
quality images or I take them personally. The site will be updated as
more images are sent my way.

The client specifically requested that the navbar be animated. I
included a few other js files such as classie and cbp.