Focus was on the navbar and the footer.  Client requested a specific style based on work from MathavanJaya
The full site has been sketched out and colored. Multiple sketches with
color were presented to the client based on our initial discussion about
the site. Different color schemes were presented but the blue, white,
and yellow were chosen.

The focus of this commit are the Header and the Footer. The header is
nearly complete with an animation missing. The Footer layout is finished
and some more social media buttons will be added once the client decides
on what other accounts will be created.

The media queries are important because the client lives in an area
where smart phones are dominant. Creating this responsive layout is the
most important to the client.